@@ -1,19 +1,20 @@
 [build-system]
-requires = ["flit_core >=2,<4"]
+requires = ["flit_core >=3.2,<4"]
 build-backend = "flit_core.buildapi"
 
 [tool.flit.metadata]
-module = "pep517"
+module = "pyproject_hooks"
+dist-name = "pyproject-hooks"
 author = "Thomas Kluyver"
 author-email = "thomas@kluyver.me.uk"
-home-page = "https://github.com/pypa/pep517"
+home-page = "https://github.com/pypa/pyproject-hooks"
 description-file = "README.rst"
 requires = [
     "tomli >=1.1.0;python_version<'3.11'",
     "importlib_metadata;python_version<'3.8'",
     "zipp;python_version<'3.8'",
 ]
-requires-python = ">=3.6"
+requires-python = ">=3.7"
 classifiers = [
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
