@@ -2,11 +2,11 @@ from os.path import abspath, dirname
 from os.path import join as pjoin
 
 import pytest
-import tomli
 from testpath import assert_isfile, modified_env
 from testpath.tempdir import TemporaryDirectory
 
-from pep517.wrappers import HookMissing, Pep517HookCaller
+from pyproject_hooks import BuildBackendHookCaller, HookMissing
+from pyproject_hooks._compat import tomllib
 
 SAMPLES_DIR = pjoin(dirname(abspath(__file__)), 'samples')
 BUILDSYS_PKGS = pjoin(SAMPLES_DIR, 'buildsys_pkgs')
@@ -15,8 +15,8 @@ BUILDSYS_PKGS = pjoin(SAMPLES_DIR, 'buildsys_pkgs')
 def get_hooks(pkg):
     source_dir = pjoin(SAMPLES_DIR, pkg)
     with open(pjoin(source_dir, 'pyproject.toml'), 'rb') as f:
-        data = tomli.load(f)
-    return Pep517HookCaller(source_dir, data['build-system']['build-backend'])
+        data = tomllib.load(f)
+    return BuildBackendHookCaller(source_dir, data['build-system']['build-backend'])
 
 
 def test_get_requires_for_build_wheel():
